more relocated modules and config omissions
