+++ scripts/config/processorconfig.py.dist
@@ -46,6 +46,8 @@ temporaryFileSystemStoragePath.default = '/home/socorro/temp'
 #---------------------------------------------------------------------------
 # local processor config
 
+from collectorconfig import dumpField
+
 elasticSearchOoidSubmissionUrl = cm.Option()
 elasticSearchOoidSubmissionUrl.doc = 'a url to submit ooids for Elastic Search (use %s in place of the ooid) (leave blank for no Elastic Search)'
 #elasticSearchOoidSubmissionUrl.default = 'http://node14.generic.metrics.sjc1.mozilla.com:9999/queue/tasks/%s'
