+++ socorro/collector/submitter.py
@@ -11,7 +11,7 @@ import signal
 
 import socorro.lib.util as sutil
 import socorro.lib.iteratorWorkerFramework as iwf
-import socorro.lib.filesystem as sfs
+import socorro.external.filesystem.filesystem as sfs
 import socorro.lib.stats as stats
 
 import poster
@@ -181,7 +181,7 @@ if __name__ == '__main__':
   except ImportError:
     import simplejson as json
   import sys
-  import socorro.lib.filesystem
+  import socorro.external.filesystem.filesystem
   import socorro.lib.uuid as uuid
 
   existingHangIdCache = {}
@@ -213,7 +213,7 @@ if __name__ == '__main__':
     print >>sys.stderr, x
 
   def walkFileSystemSubmittingReports (fileSystemRoot, serverURL, errorReporter=reportErrorToStderr, uniqueHang=False):
-    for aPath, aFileName, aJsonPathName in socorro.lib.filesystem.findFileGenerator(fileSystemRoot, lambda x: x[2].endswith("json")):
+    for aPath, aFileName, aJsonPathName in socorro.external.filesystem.filesystem.findFileGenerator(fileSystemRoot, lambda x: x[2].endswith("json")):
       try:
         dumpfilePathName = os.path.join(aPath, "%s%s" % (aFileName[:-5], ".dump"))
         submitCrashReport(aJsonPathName, dumpfilePathName, serverURL, uniqueHang)
